requirements.txt
alembic==1.12.0
bcrypt==4.0.1
bleach==6.0.0
blinker==1.6.2
boto3==1.34.74
botocore==1.34.74
click==8.1.7
exceptiongroup==1.1.3
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
packaging==23.1
pluggy==1.3.0
psycopg2==2.9.7
psycopg2-binary==2.9.8
pyclean==2.7.5
pytest==7.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.10.1
six==1.16.0
SQLAlchemy==2.0.21
tabulate==0.9.0
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.2.1
webencodings==0.5.1
Werkzeug==2.3.7
